Extracting query, mutation and subscription to the respective schema files (#3602)
