+++ litmus-portal/graphql-server/graph/gitops.graphqls
@@ -92,3 +92,35 @@ type GitConfigResponse {
     """
     sshPrivateKey: String
 }
+
+extend type Query {
+    # GIT-OPS OPERATIONS
+    """
+    Returns the git configuration for gitops
+    """
+    getGitOpsDetails(projectID: String!): GitConfigResponse! @authorized
+}
+
+extend type Mutation {
+    # GIT-OPS OPERATIONS
+    """
+    Sends workflow run request(single run workflow only) to agent on gitops notification
+    """
+    # authorized directive not required
+    gitopsNotifier(clusterInfo: ClusterIdentity!, workflowID: String!): String!
+
+    """
+    Enables gitops settings in the project
+    """
+    enableGitOps(config: GitConfig!): Boolean! @authorized
+
+    """
+    Disables gitops settings in the project
+    """
+    disableGitOps(projectID: String!): Boolean! @authorized
+
+    """
+    Updates gitops settings in the project
+    """
+    updateGitOps(config: GitConfig!): Boolean! @authorized
+}
