+++ litmus-portal/graphql-server/graph/workflow.graphqls
@@ -1,3 +1,5 @@
+directive @authorized on FIELD_DEFINITION
+
 """
 Defines the details of the weightages of each chaos experiment in the workflow
 """
@@ -561,4 +563,93 @@ type ListWorkflowsResponse {
   Details related to the workflows
   """
   workflows: [Workflow]!
+}
+
+type Query {
+  # WORKFLOW OPERATIONS
+  """
+  Returns the list of workflows in a project based on various filter parameters
+  """
+  listWorkflows(request: ListWorkflowsRequest!): ListWorkflowsResponse!
+  @authorized
+
+  """
+  Returns the list of workflow runs in a project based on various filter parameters
+  """
+  listWorkflowRuns(
+    request: ListWorkflowRunsRequest!
+  ): ListWorkflowRunsResponse! @authorized
+
+  """
+  Returns the list of predefined workflows in a project based on various filter parameters
+  """
+  listPredefinedWorkflows(hubName: String!, projectID: String!): [String!]!
+  @authorized
+
+  """
+  Returns the list of predefined experiments in a project
+  """
+  getPredefinedExperimentYAML(request: ExperimentRequest!): String! @authorized
+}
+
+type Mutation {
+  # WORKFLOW OPERATIONS
+  """
+  Creates a new workflow and applies its manifest
+  """
+  createChaosWorkFlow(request: ChaosWorkFlowRequest!): ChaosWorkFlowResponse!
+  @authorized
+
+  """
+  Reruns the workflow and applies its manifest
+  """
+  reRunChaosWorkFlow(projectID: String!, workflowID: String!): String!
+  @authorized
+
+  """
+  Updates the workflow
+  """
+  updateChaosWorkflow(request: ChaosWorkFlowRequest): ChaosWorkFlowResponse!
+  @authorized
+
+  """
+  Removes a workflow from cluster
+  """
+  deleteChaosWorkflow(
+    projectID: String!
+    workflowID: String
+    workflowRunID: String
+  ): Boolean! @authorized
+
+  """
+  Removes workflow run from the cluster only
+  """
+  terminateChaosWorkflow(
+    projectID: String!
+    workflowID: String
+    workflowRunID: String
+  ): Boolean! @authorized
+
+  """
+  Creates a new workflow run and sends it to subscriber
+  """
+  # authorized directive not required
+  chaosWorkflowRun(request: WorkflowRunRequest!): String!
+
+  """
+  Manually sync the status of the workflow run
+  """
+  syncWorkflowRun(
+    projectID: String!
+    workflowID: String!
+    workflowRunID: String!
+  ): Boolean! @authorized
+}
+
+type Subscription {
+  # WORKFLOW OPERATIONS
+  """
+  Sends workflow events to the subscriber
+  """
+  getWorkflowEvents(projectID: String!): WorkflowRun! @authorized
 }
